Add .travis.yml to support TRAVIS CI
Use TRAVIS CI to automatically build the website to gh-pages
branch every time we create a new version of it.

Right now the TRAVIS CI is configured to look only for the
fork of the user @georgegkas. In the future will have to
transfer the ownership to the upstream repository.